refactor(uniffi): split sync namespace into submodule + document password zeroize (#187)
Code-review follow-ups on Task 2: extract the sync fns/converters/tests into
namespace/sync.rs (namespace.rs was over the 500-line guideline); document why
the password move into SecretBytes needs no separate zeroize (ownership transfer
+ ZeroizeOnDrop, unlike the borrow-based wrappers); note the orphan-rule reason
the VetoDecisionDto conversion is field-by-field.
